chore: fix all scalac 2.13 warnings

+++ core/src/main/scala/com/gainstrack/core/TimeSeriesInterpolator.scala
@@ -105,8 +105,8 @@ object TimeSeriesInterpolator {
         // Linear interpolation
         //val all: N = Fractional[N].fromLong(ChronoUnit.DAYS.between(e.before._1, e.after._1))
         //val n: N = Fractional[N].fromLong(ChronoUnit.DAYS.between(e.before._1, date))
-        val all:Double = ChronoUnit.DAYS.between(e.before._1, e.after._1)
-        val n: Double = ChronoUnit.DAYS.between(e.before._1, date)
+        val all:Double = ChronoUnit.DAYS.between(e.before._1, e.after._1).toDouble
+        val n: Double = ChronoUnit.DAYS.between(e.before._1, date).toDouble
         val ratio: Double = (n / all).toDouble
         val diff: Double = (e.after._2 - e.before._2).toDouble
         val ret: Double = (diff * ratio) + e.before._2.toDouble
